+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Implementation Changelog
 
 * [X] Implemented Virtual Machine
-* [ ] Got rid of as-let-end block due to redundancy.
+* [X] Got rid of as-let-end block due to redundancy.
    Scoping is normal inside blocks. But introduced it as
    pattern-matching structure.
 * [ ] Implemented Garbage Collection